fix convert
