+++ sotopia/generation_utils/__init__.py
@@ -2,6 +2,7 @@ from .generate import (
     agenerate_env_profile,
     agenerate,
     agenerate_action,
+    agenerate_init_profile,
 )
 from .output_parsers import (
     EnvResponse,
@@ -10,6 +11,7 @@ from .output_parsers import (
     PydanticOutputParser,
     ListOfIntOutputParser,
 )
+from .convert import convert_narratives
 
 __all__ = [
     "EnvResponse",
@@ -20,4 +22,6 @@ __all__ = [
     "agenerate_env_profile",
     "agenerate",
     "agenerate_action",
+    "agenerate_init_profile",
+    "convert_narratives",
 ]
