+++ sotopia/generation_utils/convert.py
@@ -1,0 +1,37 @@
+from sotopia.generation_utils.generate import agenerate
+from sotopia.generation_utils.output_parsers import StrOutputParser
+from pydantic import validate_call
+
+
+@validate_call
+async def convert_narratives(
+    model_name: str,
+    narrative: str,
+    text: str,
+    bad_output_process_model: str | None = None,
+    use_fixed_model_version: bool = True,
+) -> str:
+    if narrative == "first":
+        return await agenerate(
+            model_name=model_name,
+            template="""Please convert the following text into a first-person narrative.
+            e.g, replace name, he, she, him, her, his, and hers with I, me, my, and mine.
+            {text}""",
+            input_values=dict(text=text),
+            output_parser=StrOutputParser(),
+            bad_output_process_model=bad_output_process_model,
+            use_fixed_model_version=use_fixed_model_version,
+        )
+    elif narrative == "second":
+        return await agenerate(
+            model_name=model_name,
+            template="""Please convert the following text into a second-person narrative.
+            e.g, replace name, he, she, him, her, his, and hers with you, your, and yours.
+            {text}""",
+            input_values=dict(text=text),
+            output_parser=StrOutputParser(),
+            bad_output_process_model=bad_output_process_model,
+            use_fixed_model_version=use_fixed_model_version,
+        )
+    else:
+        raise ValueError(f"Narrative {narrative} is not supported.")
