+++ sotopia/generation_utils/generate.py
@@ -454,40 +454,6 @@ async def agenerate_init_profile(
 
 
 @validate_call
-async def convert_narratives(
-    model_name: str,
-    narrative: str,
-    text: str,
-    bad_output_process_model: str | None = None,
-    use_fixed_model_version: bool = True,
-) -> str:
-    if narrative == "first":
-        return await agenerate(
-            model_name=model_name,
-            template="""Please convert the following text into a first-person narrative.
-            e.g, replace name, he, she, him, her, his, and hers with I, me, my, and mine.
-            {text}""",
-            input_values=dict(text=text),
-            output_parser=StrOutputParser(),
-            bad_output_process_model=bad_output_process_model,
-            use_fixed_model_version=use_fixed_model_version,
-        )
-    elif narrative == "second":
-        return await agenerate(
-            model_name=model_name,
-            template="""Please convert the following text into a second-person narrative.
-            e.g, replace name, he, she, him, her, his, and hers with you, your, and yours.
-            {text}""",
-            input_values=dict(text=text),
-            output_parser=StrOutputParser(),
-            bad_output_process_model=bad_output_process_model,
-            use_fixed_model_version=use_fixed_model_version,
-        )
-    else:
-        raise ValueError(f"Narrative {narrative} is not supported.")
-
-
-@validate_call
 async def agenerate_goal(
     model_name: str,
     background: str,
